% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHI_selection.R
\name{BHI_selection}
\alias{BHI_selection}
\title{Recommend basins by biological homogeneity index}
\usage{
BHI_selection(
  SeuratObject = NULL,
  MarkovObject = NULL,
  levels = NULL,
  prune = TRUE,
  weed = 10,
  OrgDb = "org.Mm.eg.db",
  ont = "all",
  keyType = "ENSEMBL"
)
}
\arguments{
\item{SeuratObject}{The seurat object.}

\item{MarkovObject}{The MarkovHC object.}

\item{levels}{An integer vector indicates the customized levels.}

\item{prune}{A Bloolean parameter indicates wether to merge small basins into
big basins. Default is TRUE.}

\item{weed}{An integer value defines small basins. When the prune parameter is
set to TRUE, the basins include less points than weed will be merged into
big basins on each level.}

\item{OrgDb}{OrgDb, and this parameter is passed to 'enrichGO' function in 'clusterProfiler' package.}

\item{ont}{One of "MF", "BP", and "CC" subontologies. And this parameter is passed to 'enrichGO' function in 'clusterProfiler' package.}

\item{keyType}{keytype of input gene, and this parameter is passed to 'enrichGO' function in 'clusterProfiler' package.}
}
\value{
A BHI vector with the names of basins is returned. This function creates a 'GOEachBasin' folder in the current directory and saves the
gene ontology enrichment result of each basin in it.
}
\description{
Function \code{BHI_selection} The function recommends basins
by biological homogeneity index.
}
\details{
This function calculate the biological homogeneity index(BHI) of basins on the customized levels. The definition of
BHI is refered to 'clValid' package. And the BHI is in the range [0, 1], with larger values corresponding to more biologically homogeneous basins.

To calculate the BHI of a basin, the differentially expressed genes among its sub-basins are found and the BHI is calculated according
to the denifition in 'clValid' package. Details are in 'MarkovHC' paper.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
