% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyGap_selection.R
\name{energyGap_selection}
\alias{energyGap_selection}
\title{Recommend levels with possible biological meaning}
\usage{
energyGap_selection(MarkovObject = NULL, m = 3)
}
\arguments{
\item{MarkovObject}{The output of the function, \code{MarkovHC}.}

\item{m}{An integer value. A (local) peak is defined as
a point such that m points either side of it has a lower or equal value to it.}
}
\value{
This function plots the energy gap along the increasing levels,
red points indicate recommend levels with possible biological meaning,
and a red circle highlight the level that may with an optimal cluster number.

A list consist of two components are returned. p is a vector contains the recommend
levels with possible biological meaning. recommend_level is the level that may
with an optimal cluster number.
}
\description{
Function \code{energyGap_selection} The function recommends levels
with possible biological meaning.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
