% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLabels.R
\name{fetchLabels}
\alias{fetchLabels}
\title{Extract the clustering labels}
\usage{
fetchLabels(MarkovObject = NULL, MarkovLevels = NULL, prune = TRUE, weed = 10)
}
\arguments{
\item{MarkovObject}{The output of the function, \code{MarkovHC}.}

\item{MarkovLevels}{An integer vector indicates the customized levels.}

\item{prune}{A Bloolean parameter indicates wether to merge small basins into
big basins. Default is TRUE.}

\item{weed}{An integer value defines small basins. When the prune parameter is
set to TRUE, the basins include less points than weed will be merged into
big basins on each level.}
}
\value{
This function returns a dataframe, whose columns correspond to customized levels,
rows correspond to samples, and entries are clustering labels.
}
\description{
Function \code{fetchLabels} The function extracts the clustering results
of each sample on customized levels.
}
\details{
This function extracts the clustering results of each sample on customized levels
and returns a dataframe.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
