% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTransitionPath.R
\name{findTransitionPath}
\alias{findTransitionPath}
\title{Find the transition path from basinA to basinB}
\usage{
findTransitionPath(
  MarkovObject = NULL,
  level = NULL,
  basinA = NULL,
  basinB = NULL
)
}
\arguments{
\item{MarkovObject}{The output of the function, \code{MarkovHC}.}

\item{level}{An integer value indicates the customized level.}

\item{basinA}{An integer value indicates the customized "from" basin.}

\item{basinB}{An integer value indicates the cusomized "to" basin.}
}
\value{
This function returns a list consists of three components:
label is a vector, the indices of 1 in this vector indicate the samples on the transition
path.

pathVertex is the vector contains indices of base-clusters on the transition path. If the
dobasecluster parameter of \code{MarkovHC} was set to FALSE, pathVertex contains the sample
indices. If the dobasecluster parameter in \code{MarkovHC} was set to TRUE, pathVertex
contains the indices of clusters on level 1.

shortestPathLength is pseudo energy cost from basinA to basinB.
}
\description{
Function \code{findTransitionPath} The function finds the samples on the
transition path from basinA to basinB.
}
\details{
The function finds the samples on the transition path from basinA to basinB on
the customized level.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
