% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMI_selection.R
\name{IMI_selection}
\alias{IMI_selection}
\title{Recommend levels by integrating internal measures}
\usage{
IMI_selection(MarkovObject = NULL, prune = NULL, weed = NULL)
}
\arguments{
\item{MarkovObject}{The output of the function, \code{MarkovHC}.}

\item{prune}{A Bloolean parameter indicates wether to merge small basins into
big basins. Default is TRUE.}

\item{weed}{An integer value defines small basins. When the prune parameter is
set to TRUE, the basins include less points than weed will be merged into
big basins on each level.}
}
\value{
This function returns a dataframe containing the ranked levels and internal measures.
}
\description{
Function \code{IMI_selection} The function recommends levels
by integrating internal measures.
}
\details{
IMI_selection deploys connectivity, silhouette and Dunn from "clValid" package.
And aggregate them with energy gaps in MarkovObject to rank the levels.

Refer to the manuscript of "clValid" package, the connectivity has a value between zero and
infinite and should be minimized, well-clustered observations having silhouette values near 1
and poorly clustered observations having silhouette values near 1, and the Dunn Index has a value
between zero and infinite, and should be maximized. Meanwhile, the C_cut value of each level in MarkovHC
indicates the energy level, so the first order difference of the C_cut values, which we termed energy gaps(C_cut_gap)
indicates the energy difference between neighbor levels and this value should be maximized. We rank levels
by these four measures and aggrate the ranks using aggregateRanks function in "RobustRankAggreg" package.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
