% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_RS.R
\name{judge_RS}
\alias{judge_RS}
\title{Get a vector to judge recurrent state}
\usage{
judge_RS(P = NULL)
}
\arguments{
\item{P}{A transition matrix.}
}
\value{
\tabular{ll}{
    \code{RS} \tab A vector used to judge recurrency. If the i th entry is non-zero, then state i is a recurrent state. \cr
    }
}
\description{
Function '\code{judge_RS}' gets a vector helping to judge recurrent state based on a transition matrix.
}
\details{
Assume that we have an input transition matrix P. We can first find out all
eigen values and eigen vectors of P^T. As we know that if a state i is a recurrent state,
then there must be a eigen value 1 with a corresponding eigen vector whose i th entry
is non-zero. If a state i is a transient state, then for all eigen values 1, the i th entry
of the corresponding eigen vector must be 0. According to the above conclusions, we can easily
judge whether a state is recurrent or not.
}
\examples{
P<-matrix(c(1,0,0,0,0.5,0,0,0.5,1),nrow=3)
RS<-judge_RS(P)
}
\author{
Zhenyi Wang
}
