% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_selection.R
\name{PC_selection}
\alias{PC_selection}
\title{Select top principle components}
\usage{
PC_selection(SeuratObject = NULL)
}
\arguments{
\item{SeuratObject}{A seurat object after RunPCA run.}
}
\value{
This function plots the standard deviations of the principle components and returns
the location(change point)+1.
}
\description{
Function \code{PC_selection} The function selects top the principle components by
finding the change point of standard deviations of each principle component.
}
\details{
The piecewise.linear function in "SiZer" package is used to fit a degree 1 spline
with 1 change point and return the location(change point)+1.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
