% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHierarchicalStructure.R
\name{plotHierarchicalStructure}
\alias{plotHierarchicalStructure}
\title{Plot the MarkovHC hierarchical structure or return the tree}
\usage{
plotHierarchicalStructure(
  MarkovObject = NULL,
  MarkovLevels = NULL,
  colorVector = NULL,
  plot = TRUE,
  prune = TRUE,
  weed = 10
)
}
\arguments{
\item{MarkovObject}{The output of the function, \code{MarkovHC}.}

\item{MarkovLevels}{An integer vector indicates the customized levels.}

\item{colorVector}{A character vector set the colors of basins on the first customized level.
The length of this vector must equal to the number of basins on the first customized level.}

\item{plot}{A Bloolean parameter indicates wether to plot the MarkovHC hierarchical structure.
If this parameter was set to FALSE, this function will not plot and a phylo tree will be returned. The
phylo tree could be plotted by some tree visualization packages, such as "ggtree".}

\item{prune}{A Bloolean parameter indicates wether to merge small basins into
big basins. Default is TRUE.}

\item{weed}{An integer value defines small basins. When the prune parameter is
set to TRUE, the basins include less points than weed will be merged into
big basins on each level. Default is 10.}
}
\value{
When the plot parameter is set to TRUE, this function plots the MarkovHC hierarchical structure. When the plot
parameter is set to FALSE, this function returns a phylo tree and it could be plotted by other tree visualization
packages, such as "ggtree".
}
\description{
Function \code{plotHierarchicalStructure} The function plots the MarkovHC hierarchical structure or returns the tree.
}
\details{
This function plots the MarkovHC hierarchical structure or returns the tree.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
