% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepWisepath.R
\name{stepWisepath}
\alias{stepWisepath}
\title{Extract stepwise path points and transition points}
\usage{
stepWisepath(MarkovObject = NULL, MarkovLevel = NULL, stepBasin = NULL)
}
\arguments{
\item{MarkovObject}{The output of the function, \code{MarkovHC}.}

\item{MarkovLevel}{An integer value indicates the customized level.}

\item{stepBasin}{An integer vector indicates the customized basins on the transition path.}
}
\value{
This function returns a list consists of two components:
MarkovHCpathPoint is a vector contains the indices of samples along the transition path.

transitionPoint is a list contains the indices of samples identified as critical points
along the transition path.
}
\description{
Function \code{stepWisepath} The function finds the samples on the
stepwise transition path.
}
\details{
This function finds the samples on the transition path along the customized basins
in stepBasin parameter.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
