% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovHC.R
\name{MarkovHC}
\alias{MarkovHC}
\title{Markov Hierarchical Clustering Algorithm}
\usage{
MarkovHC(
  MarkovHC_input = NULL,
  SNNslot = NULL,
  KNNslot = NULL,
  KNN = 20,
  dobasecluster = TRUE,
  cutpoint = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{MarkovHC_input}{A Seurat object (recommend) or a matrix/dataframe with row names and column names.
Each row should be a feature and each column should be a sample.}

\item{SNNslot}{A character vector indicates the name of the SNN graph slot in the seurat object. If the MarkovHC_input
is a matrix, this parameter will be ignored.}

\item{KNNslot}{A character vector indicates the name of the KNN graph slot in the seurat object. If the MarkovHC_input
is a matrix, this parameter will be ignored.}

\item{KNN}{An interger indicates the number of neighbors in building the
KNN graph. Default is 20. If the MarkovHC_input is a Seurat object, the KNN
parameter will be set as that in the Seurat object.}

\item{dobasecluster}{A Bloolean parameter indicates wether to do Louvain clustering
on the first level. Default is TRUE.}

\item{cutpoint}{A numeric value in [0,1] indicates the threshold of
the quantile of pseudo energy, default is 0.001.}

\item{verbose}{A Bloolean parameter indicates wether to print details when run
the program.}
}
\value{
This function returns a list including three components, hierarchicalStructure, inputParameters,
and midResults, which are the hierarchical clustering results, the input parameters, and some intermediate results.
}
\description{
Function \code{MarkovHC} The main function of MarkovHC, which applies Markov
Hierarchical Clustering Algorithm to the given data (a Seurat object or a
matrix/dataframe).
}
\details{
The data given by \code{MarkovHC_input} will be clustered by using
the Markov Hierarchical Clustering Algorithm, which generates a hierarchical
structure based on the metastability of exponentially perturbed Markov chain.
More details of the algorithm please refer to the article.
}
\author{
Zhenyi Wang wangzy17@mails.tsinghua.edu.cn
}
